Symmetry hunt: corpus minor asymmetries 7 -> 1 - king proximity and bishop-pawns accumulators upgraded float->double (same class as the long-term mobility fix: square-order summation differs between mirrored boards, float epsilon flips the final int truncation). Last remaining case is a +-1cp pawn-structure TOTAL whose hidden float sum rounds differently from its displayed integer components - documented, telemetry keeps counting, tolerance covers it. Corpus: 2170 positions, 1 minor asymmetry

+++ opti_chess/eval_pawn.cpp
@@ -869,7 +869,7 @@ int Board::get_bishop_pawns() const {
 	//cout << "white_pawns_blocked: " << white_central_pawns_blocked << endl;
 	//cout << "black_pawns_blocked: " << black_central_pawns_blocked << endl;
 
-	float bishop_pawns_value = 0.0f;
+	double bishop_pawns_value = 0.0;
 
 	for (uint8_t row = 0; row < 8; row++) {
 		for (uint8_t col = 0; col < 8; col++) {
